WFP-998 added currently retrieved data mapped to email template (#88)
* WFP-998 added currently retrieved data mapped to email template

* WFP-998 added getting more of the data for email

* WFP-998 added calculating induction date for email

* WFP-998 changed conviction id to big integer

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsworkload/domain/ImpactCase.kt
@@ -1,8 +1,9 @@
 package uk.gov.justice.digital.hmpps.hmppsworkload.domain
 
 import com.fasterxml.jackson.annotation.JsonCreator
+import java.math.BigInteger
 
 data class ImpactCase @JsonCreator constructor(
   val crn: String,
-  val convictionId: Long
+  val convictionId: BigInteger
 )
